// src/app/services/models/student-dto-to-send.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface StudentDtoToSend {
  birthday?: string;
  email?: string;
  firstName?: string;
  lastName?: string;
  password?: string;
  phone?: string;
  schoolYear?: string;
  sexe?: 'Men' | 'Women';
  image?:File;
}