// src/app/services/models/professor-to-send.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ProfessorToSend {
  birthday?: string;
  departement?: 'BackendDevelopemet' | 'FullStackDevelopement' | 'FrontDevelopement' | 'AI_MachineLearning';
  email?: string;
  firstName?: string;
  lastName?: string;
  password?: string;
  phone?: string;
  sexe?: 'Men' | 'Women';
  image?:File;
}